+++ packages/plugins/document/index.ts
@@ -1,0 +1,2 @@
+export { LocalResearchReportInputResolver, PdfJsResearchReportTextExtractor } from './research-report-input-resolver.ts'
+export type { LocalResearchReportInputResolverOptions, ResearchReportPdfTextExtractor } from './research-report-input-resolver.ts'